test: force init argilla client credentials (#3229)
# Description

This PR fix the weird failing test by forcing an Argilla init with
default user credentials before the test operations.

Closes #3228

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [x] Test fix

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

The test works now.