docs: fixed typo described in #3238 (#3240)
<!-- Thanks for your contribution! As part of our Community Growers
initiative 🌱, we're donating Justdiggit bunds in your name to reforest
sub-Saharan Africa. To claim your Community Growers certificate, please
contact David Berenstein in our Slack community or fill in this form
https://tally.so/r/n9XrxK once your PR has been merged. -->

# Description

I worked ion this and resolved some typos

Closes #3238 

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [X] Documentation update

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

N.A.

**Checklist**

- [X] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)