docs: formatting of libraries on tutorial main page as described in #2819 (#3255)
<!-- Thanks for your contribution! As part of our Community Growers
initiative 🌱, we're donating Justdiggit bunds in your name to reforest
sub-Saharan Africa. To claim your Community Growers certificate, please
contact David Berenstein in our Slack community or fill in this form
https://tally.so/r/n9XrxK once your PR has been merged. -->

# Description

Please include a summary of the changes and the related issue. Please
also include relevant motivation and context. List any dependencies that
are required for this change.

I fixed the formatting of the libraries on the main tutorial page, so
that it previews all 21 different options. Additionally, there were two
references to OpenAI in the sidebar and on the libraries page, so I
fixed that as well.

Closes #2819

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)
- [X] Documentation update

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

N.A.

**Checklist**

- [X] I have merged the original branch into my forked branch
- [X] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)