docs: fixed styling of code output blocks as described in #2707 (#3270)
<!-- Thanks for your contribution! As part of our Community Growers
initiative 🌱, we're donating Justdiggit bunds in your name to reforest
sub-Saharan Africa. To claim your Community Growers certificate, please
contact David Berenstein in our Slack community or fill in this form
https://tally.so/r/n9XrxK once your PR has been merged. -->

# Description

Please include a summary of the changes and the related issue. Please
also include relevant motivation and context. List any dependencies that
are required for this change.

For output blocks that use the standard error styling (`div.nboutput`
... `stderr`), I set the font to stay at gray/black even for dark mode
users, so the text remains visible (unlike the white text on the pink
background like before) in `config.py`.

Closes #2707 

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [X] Documentation update

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

N.A.

**Checklist**
- [X] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)


Note: Please ignore all of the other 11 commits, only the ones with
#2707 in the commit message are relevant!

---------

Co-authored-by: David Berenstein <david.m.berenstein@gmail.com>