feat: migrate to async SQLAlchemy engine (#3162)
# Description

This PR updates the API to use an async SQLAlchemy engine to connect to
the DB:

- Added a new `AsyncEngine` instance, `sessionmaker` that will create
`AsyncSession`s and a new async generator function `get_async_db` that
will be used in the API. The sync engine and `sessionmaker` creating
`Session`s are still there as we need them for the CLI.
- Update the `Base` (`DeclarativeBase`) SQLAlchemy ORM class to also
inherit from `AsyncAttrs` class. This will allow to load/get
relationships using `relationship = await
model.awaitable_attrs.relationship_attribute_name`. When using the sync
engine, this was not required because of the lazy loading, but using the
async engine [this is not
possible](https://docs.sqlalchemy.org/en/20/orm/extensions/asyncio.html#asyncio-orm-avoid-lazyloads).
- Update all the `PolicyAction`s methods/function to be async (required
because they are used from the API and the API will use the async
connection to the DB)
- Added a new `AsyncSQLAlchemyModelFactory` factory_boy factory that
allows to create SQLAlchemy ORM models using an async connection. All
the factories classes have been updated to use this class as parent
class. In some cases, like for the `UserFactory` which is used in some
tests for the CLI which uses the sync DB connection, a sync version of
the factory has been created (`UserSyncFactory`).
- Update parts of the code in which we were adding, updating or deleting
records for a Feedback Dataset to rollback the DB transaction (i.e. not
doing any change in the DB) if an exception is raised while inserting
the record info in ElasticSearch.

Closes #3035 

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [x] New feature (non-breaking change which adds functionality)
- [x] Refactor (change restructuring the codebase without changing
functionality)
- [x] Improvement (change adding some improvement to an existing
functionality)

**How Has This Been Tested**

All the tests using the async connection to the DB have been updated to
be async. The previous unit tests that not require to connect to the DB
(at least directly) are passing.

**Checklist**

- [x] I have merged the original branch into my forked branch
- [ ] I added relevant documentation
- [x] follows the style guidelines of this project
- [x] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)