fix: unit test failing if not local db (#3307)
# Description

The API startup event was using the
`argilla.server.database.get_async_db` function to get a DB session to
check if the default argilla user exists. We were overriding the
`get_async_db` dependency for the API endpoints, but we were not mocking
the function when being used from the API startup, and therefore the
function was returning a session to the local sqlite DB
(`$HOME/.argilla/argilla.db`). If the sqlite local db file didn't exist
and the migrations hadn't been applied, then a `sqlalchemy.errors.
OperationalError` was being raised as the `User` table didn't exist.

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Refactor (change restructuring the codebase without changing
functionality)
- [ ] Improvement (change adding some improvement to an existing
functionality)
- [ ] Documentation update

**How Has This Been Tested**

Running the tests without the `$HOME/.argilla/argilla.db` file doesn't
raise an exception anymore.

**Checklist**

- [ ] I added relevant documentation
- [x] follows the style guidelines of this project
- [x] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [x] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)

---------

Co-authored-by: frascuchon <francis@argilla.io>