feat: add `Suggestion` endpoints (#3304)
# Description

This PR adds the following endpoints:

- `GET /api/v1/records/{record_id}/suggestions`
- `POST /api/v1/records/{record_id}/suggestions`

and updates the following ones:

- `GET /api/v1/datasets/{dataset_id}/records`: output payload updated to
add suggestions depending on the value of the query parameter `include`
- `GET /api/v1/me/datasets/{dataset_id}/records`: output payload updated
to add suggestions depending on the value of the query parameter
`include`
- `POST /api/v1/me/datasets/{dataset_id}/records/search`: output payload
updated to add suggestions depending on the value of the query parameter
`include`
- `POST /api/v1/datasets/{dataset_id}/records`: input payload updated to
add suggestions

Additionally and not directly related to the suggestions, this PR has
updated the `GET /api/v1/me/datasets/{dataset_id}/records` endpoint to
only include the responses of the record in the output payload when the
value `responses` has been provided via the `include` query parameter.
Old behaviour always returned the responses.

Finally, I've added a `TimestampMixin` class that can be used with a
SQLAlchemy ORM model class to add the `inserted_at` and `updated_at`
attributes.

Some notes about the suggestions:

- I've decided that it is a good idea to leave the `Suggestion.score`
value unrestricted. At first I thought it would be good to force the
user to provide a value between `0` and `1`, but users may want to
assign a suggestion based on a distance based metric whose value is not
bounded between `0` and `1`.
- `Suggestion.type` has only two possible values for now: `human` and
`model` (machine learning, rule-based, heuristic, etc). The idea is to
complement with `Suggestion.agent`.

Closes #3277

**Type of change**

- [x] Refactor
- [x] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)

**How Has This Been Tested**

Calling the endpoints in a local environment manually and adding unit
tests covering all the possible cases.

**Checklist**

- [ ] I added relevant documentation
- [x] follows the style guidelines of this project
- [x] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [x] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)