Docs/3350 docs not all extra package dependencies are installed (#3351)
<!-- Thanks for your contribution! As part of our Community Growers
initiative 🌱, we're donating Justdiggit bunds in your name to reforest
sub-Saharan Africa. To claim your Community Growers certificate, please
contact David Berenstein in our Slack community or fill in this form
https://tally.so/r/n9XrxK once your PR has been merged. -->

# Description

We were building the docs.

Closes #3350 

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [X] Bug fix (non-breaking change which fixes an issue)
- [X] Documentation update

**How Has This Been Tested**
N.A.
**Checklist**

- [X] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)