chore: added telemetry to ArgillaTrainer (#3336)
<!-- Thanks for your contribution! As part of our Community Growers
initiative 🌱, we're donating Justdiggit bunds in your name to reforest
sub-Saharan Africa. To claim your Community Growers certificate, please
contact David Berenstein in our Slack community or fill in this form
https://tally.so/r/n9XrxK once your PR has been merged. -->

# Description

Please include a summary of the changes and the related issue. Please
also include relevant motivation and context. List any dependencies that
are required for this change.

Closes #3325 

**Type of change**

- [X] New feature (non-breaking change which adds functionality)

**How Has This Been Tested**

N.A.

**Checklist**

N.A.