feat: update create suggestion endpoint to `PUT` (#3391)
# Description

This PR updates the `POST` create suggestion endpoint added in #3304 to
be a `PUT` endpoint instead that allows creating a suggestion for a
given combination of question + record (if it doesn't exist) or update
an existing one:

- The endpoint will return a `201` when the suggestion is created
- The endpoint will return a `200` when the suggestion existed and it
has been updated.

Also, this PR updates the `Suggestion` SQLAlchemy ORM model adding a new
unique constraint for record id and question id. As the migration file
added in #3304 to create the `suggestions` table has not been applied
yet in any prod environment, just in `dev.argilla.io`, I modified the
migration file to include this unique constraint. Before deploying these
changes in `dev.argilla.io` we will need to:

1. Revert the migration
2. Apply the latest migration

**Type of change**

- [x] New feature (non-breaking change which adds functionality)
- [x] Breaking change

**How Has This Been Tested**

Manually in a local environment and I've added unit tests.

**Checklist**

- [ ] I added relevant documentation
- [x] follows the style guidelines of this project
- [x] I did a self-review of my code
- [x] I made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [x] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)