fix: add `DeprecatedDatasetConfig` for backwards compatibility & improve `warnings` (#3416)
<!-- Thanks for your contribution! As part of our Community Growers
initiative 🌱, we're donating Justdiggit bunds in your name to reforest
sub-Saharan Africa. To claim your Community Growers certificate, please
contact David Berenstein in our Slack community or fill in this form
https://tally.so/r/n9XrxK once your PR has been merged. -->

# Description

This PR addresses the bug fix reported by @tomaarsen at #3414, so that
we use `DeprecatedDatasetConfig` instead of `DatasetConfig` and to make
`from_json` a `classmethod`, to make sure that previously uploaded
`FeedbackDataset`s to the HuggingFace Hub can still be loaded with
`FeedbackDataset.from_huggingface`.

Additionally, as reported by @nataliaElv the warning messages when
adding `suggestions` to a `FeedbackRecord` could be improved, so we've
kept just the necessary and adding an `Ignore this exception ...` in
cases where warning is not needed.

Closes #3414

**Type of change**

- [X] Bug fix (non-breaking change which fixes an issue)

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

- [X] Run
`FeedbackDataset.from_huggingface("argilla/stackoverflow_feedback_demo")
- [X] Update unit tests to catch updated warnings

**Checklist**

- [ ] I added relevant documentation
- [X] follows the style guidelines of this project
- [X] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [X] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [ ] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)

---------

Co-authored-by: Gabriel Martin <gabriel@argilla.io>