feat: update `package.yml` triggers (#3422)
# Description

This PR updates the triggers of the `package.yml` workflow:

- Remove all the branches but `main`, `develop` and `releases/**` from
the `push` trigger
- Add `workflow_dispatch` trigger to allow triggering the `package.yml`
workflow manually.

This way, the unit tests will gets executed just one time per push when
a PR for the branch is opened. If there isn't a PR opened for the branch
and for some reason the Docker image for that branch needs to be build,
then the `package.yml` workflow can get executed manually.

This PR is an option and a personal opinion, if you agree we can merge
it. Otherwise, I'll close the PR.

**Type of change**

- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] Refactor (change restructuring the codebase without changing
functionality)
- [x] Improvement (change adding some improvement to an existing
functionality)

**How Has This Been Tested**

**Checklist**

- [ ] I added relevant documentation
- [x] follows the style guidelines of this project
- [x] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [ ] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)