feat: create user with workspaces (#3462)
# Description

This PR updates the API and the Python client so a list of workspaces
names to which the user should be linked to can be provided. This is
useful when creating a new user, as just one request will be needed,
instead of the current behaviour, which requires one to create the user
and another one to add the user to the workspace.

In addition, the code blocks in which a commit or a rollback has to be
done depending on wether an exception is raised have been updated to use
the async context manager `async with session.begin_nested():` which
automatically will commit if the block is executed without any problem
or rollback if an exception is raised.

**Type of change**

- [x] New feature (non-breaking change which adds functionality)
- [x] Improvement (change adding some improvement to an existing
functionality)

**How Has This Been Tested**

I've been able to create a user with workspaces in a local development
environment. Also, the unit tests regarding the user creation have been
updated.

**Checklist**

- [ ] I added relevant documentation
- [x] follows the style guidelines of this project
- [x] I did a self-review of my code
- [x] I made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [x] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)