tests:  support for running server tests with postgresql database (#3487)
<!-- Thanks for your contribution! As part of our Community Growers
initiative 🌱, we're donating Justdiggit bunds in your name to reforest
sub-Saharan Africa. To claim your Community Growers certificate, please
contact David Berenstein in our Slack community or fill in this form
https://tally.so/r/n9XrxK once your PR has been merged. -->

# Description

The main target of this PR is to provide support for running server
tests using a PostgreSQL database. This may help to find different
behavior depending on which dataset was used.

Since Argilla is using an async connection for the database and the
standard `TestClient` client is not fully compatible when using other
than SQLite async connections in tests (see
[here](https://fastapi.tiangolo.com/advanced/async-tests/#httpx) for
more details), the tests have been split into 2 groups: `unit` and
`integration`

The first group will affect to tests checking single component
functionality, and the `httpx.AsyncClient` will be used for those
server-related tests.

The second group (integration) will check a more complex workflow, where
several components may be affected. For now, those tests are using a
mocked variant of the fastapi `TestClient` but ideally, this will change
in the future to use a real connection to an up-and-running Argilla
server.


Also, this new test organization allows us to optimize the build
pipeline by only launching the unit tests using different search backend
versions, which reduces the execution time.

Closes #3291

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [X] Refactor (change restructuring the codebase without changing
functionality)
- [X] Improvement (change adding some improvement to an existing
functionality)

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

Several tests with multiple configurations have been launched locally 

**Checklist**

- [X] follows the style guidelines of this project
- [X] I did a self-review of my code
- [X] I made corresponding changes to the documentation

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Gabriel Martín Blázquez <gmartinbdev@gmail.com>
Co-authored-by: Tom Aarsen <37621491+tomaarsen@users.noreply.github.com>