fix: cannot update ranking & rating question with `settings` (#3552)
# Description

The input payload for updating a question was not consistent across the
different types of questions. For the `text`, `label_selection` and
`multi_label_selection`, it is necessary to provide the kind of question
to be updated including `{"settings": {"type": "question_type"}` in the
input payload, while for the `rating` and `ranking` question it was not
needed to provide the `settings` and when provided it caused a
validation error. This was because this last two questions types didn't
have a `pydantic.BaseModel` included in the `QuestionUpdateSettings`
discriminator causing the following payloads to not be valid when
updating them:

```json
{
   "name": "New Name",
    "settings": {
        "type": "ranking"
    }
}
```

and the only valid payload was:

```json
{
   "name": "New Name"
}
```

To keep the consistency in the input payload for updating a question,
this PR include two new `pydantic.BaseModel`, one for rating and one for
ranking question, that are included in the `QuestionUpdateSettings`
discriminator so the payload mentioned above is valid when updating a
`ranking` or `rating` question.

**Type of change**

- [x] Bug fix (non-breaking change which fixes an issue)

**How Has This Been Tested**

Unit tests has been updated and manual tests on local environment.

**Checklist**

- [x] I followed the style guidelines of this project
- [x] I did a self-review of my code
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [x] I have added relevant notes to the `CHANGELOG.md` file (See
https://keepachangelog.com/)