fix: response with `draft` status was returning `422` (#3623)
# Description

This PR fix an issue in which the server returned a `422` when trying to
update the response of a record with `draft` status.

In addition, this PR updates `check_response` method used to validate
the responses received for a question to also received the `status` of
the response. This allows to execute partial validations of the
responses depending on the received status. For example, for the
`RankingQuestion` we only want to validate that all the possible options
have been ranked if the status is `submitted`, otherwise, we just want
to validate that the options that have been ranked are valid:

Closes #3612

**Type of change**

- [x] New feature
- [x] Bug fix (non-breaking change which fixes an issue)

**How Has This Been Tested**

In a local development environment:

- [x] Record response can be updated with `draft` status

**Checklist**

- [x] I followed the style guidelines of this project
- [x] I did a self-review of my code
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [x] I have added relevant notes to the `CHANGELOG.md` file (See
https://keepachangelog.com/)