fix: `users create` command was not printing workspace names (#3713)
# Description

In #3667 a new command to create users was added. This command prints a
summary of the created user which includes a list of workspaces to which
the users belongs. Instead of the name of the workspaces, workspaces
classes were being printed. This PR fixes this.

**Type of change**

- [x] Bug fix (non-breaking change which fixes an issue)

**How Has This Been Tested**

In a local development environment:

- [x] `python -m argilla users create --username gabriel --role
annotator --workspace ws1 --workspace ws2`: prints `Workspaces: ws1,
ws2`

**Checklist**

- [x] I followed the style guidelines of this project
- [x] I did a self-review of my code
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [ ] I have added relevant notes to the `CHANGELOG.md` file (See
https://keepachangelog.com/)