refactor(engine): define a base search engine (#3750)
# Description

This PR adds a refactor in the `search_engine` in which a base class is
defined and all implementation details are moved to the
`OpenSearchEngine` implementation.

This will be the base to extend and integrate different search engines
in Argilla.

Also, some minor changes are done in the search engine data models that
affect some API schema definitions.

Refs #3028 

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [X] Refactor (change restructuring the codebase without changing
functionality)

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

Everything is working as expected in a local deployment

**Checklist**

- [ ] I added relevant documentation
- [x] follows the style guidelines of this project
- [x] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [ ] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>