fix: use `Enum.value` when building f-strings
This was leading Argilla to fail in Python 3.11 as accessing the `Enum` fields will no longer print-out the value but the `Enum`'s reference to the value, instead we should access the `value` attribute which is available in every Python version. P.S. it still works when comparing it with a string, using it as dict key, etc.